+++ test/ui/ui-e2e.test.ts
@@ -62,6 +62,9 @@ async function testWeb(page: Page, input: string, options_: Options = {}) {
 	);
 	await setCheckbox(page, "dropAllCaps", !!options.dropAllCaps);
 	await setCheckbox(page, "escape", !!options.escape);
+	if (options.escape) {
+		await setCheckbox(page, "escapeLegacy", options.escape === true);
+	}
 	await setCheckbox(page, "stripComments", !!options.stripComments);
 	await setCheckbox(page, "encodeUrls", !!options.encodeUrls);
 	await setCheckbox(page, "tidyComments", !!options.tidyComments);